fixed vote endpoints and updated urls
