+++ contest_api/serializers/photo/retrieve_detail.py
@@ -1,6 +1,7 @@
 from rest_framework import serializers
 
 class PhotoDetailSerializer(serializers.Serializer):
+    id = serializers.IntegerField()
     title = serializers.CharField(max_length=100)
     description = serializers.CharField(max_length=300)
     pub_date = serializers.DateField()
