+++ contest_api/services/vote/create.py
@@ -5,11 +5,11 @@ from models_app.models import Photo, User, Vote
 from django import forms
 
 class CreateVoteService(ServiceWithResult):
-    photo_id = forms.IntegerField()
+    id = forms.IntegerField()
     user = ModelField(User)
 
     def process(self): #pyright: ignore
-        photo_id = self.cleaned_data.get('photo_id')
+        photo_id = self.cleaned_data.get('id')
         user = self.cleaned_data.get('user')
 
         try:
