+++ contest_api/views/vote/views.py
@@ -1,0 +1,28 @@
+from rest_framework.views import APIView
+from rest_framework.permissions import IsAuthenticated
+from rest_framework.response import Response
+from service_objects.services import  ServiceOutcome
+from contest_api.services import CreateVoteService,DeleteVoteService
+
+class VoteAPIView(APIView):
+    permission_classes = [IsAuthenticated]
+
+    def post(self,request, *arg, **kwargs):
+        outcome: ServiceOutcome = ServiceOutcome(
+            CreateVoteService,
+            {
+                **kwargs, 'user': request.user
+            }
+        )
+        
+        return Response(outcome.response_status)
+
+    def delete(self, request, *args, **kwargs):
+        outcome: ServiceOutcome = ServiceOutcome(
+            DeleteVoteService,
+            {
+                **kwargs, 'user': request.user
+            }
+        )
+        return Response(outcome.response_status)
+
